--- packages/compiler/src/transform/templates-and-concat/__tests__/flattenExpressionToParts.test.ts
@@ -1,176 +1,0 @@
-import { describe, it, expect } from 'vitest';
-import * as parser from '@babel/parser';
-import traverse, { NodePath } from '@babel/traverse';
-import * as t from '@babel/types';
-import { flattenExpressionToParts } from '../flattenExpressionToParts';
-import { mergeAdjacentStaticParts } from '../mergeAdjacentStaticParts';
-import { buildTransformResult } from '../buildTransformationResult';
-/**
- * Parse code and extract the first expression's NodePath, then run the callback.
- */
-function withExpressionPath(
-  code: string,
-  callback: (path: NodePath<t.Expression>) => void
-) {
-  const ast = parser.parse(code, {
-    sourceType: 'module',
-    plugins: ['typescript'],
-  });
-  traverse(ast, {
-    ExpressionStatement(path) {
-      callback(path.get('expression') as NodePath<t.Expression>);
-      path.stop();
-    },
-  });
-}
-
-describe('flattenExpressionToParts', () => {
-  it('flattens a string literal', () => {
-    withExpressionPath('"hello"', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toEqual([{ type: 'static', value: 'hello' }]);
-    });
-  });
-
-  it('flattens a numeric literal', () => {
-    withExpressionPath('42', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toEqual([{ type: 'static', value: '42' }]);
-    });
-  });
-
-  it('flattens a boolean literal', () => {
-    withExpressionPath('true', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toEqual([{ type: 'static', value: 'true' }]);
-    });
-  });
-
-  it('flattens null', () => {
-    withExpressionPath('null', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toEqual([{ type: 'static', value: 'null' }]);
-    });
-  });
-
-  it('flattens a plain template literal', () => {
-    withExpressionPath('`hello`', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toEqual([{ type: 'static', value: 'hello' }]);
-    });
-  });
-
-  it('flattens "A" + "B" into two static parts', () => {
-    withExpressionPath('"A" + "B"', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toEqual([
-        { type: 'static', value: 'A' },
-        { type: 'static', value: 'B' },
-      ]);
-    });
-  });
-
-  it('flattens template with expression', () => {
-    withExpressionPath('`A${name}B`', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toHaveLength(3);
-      expect(parts[0]).toEqual({ type: 'static', value: 'A' });
-      expect(parts[1].type).toBe('dynamic');
-      expect(parts[2]).toEqual({ type: 'static', value: 'B' });
-    });
-  });
-
-  it('flattens an identifier as dynamic', () => {
-    withExpressionPath('name', (path) => {
-      const parts = flattenExpressionToParts(path.node, path);
-      expect(parts).toHaveLength(1);
-      expect(parts[0].type).toBe('dynamic');
-    });
-  });
-
-  it('recognizes derive() as a derive part', () => {
-    const code = `import { derive } from 'gt-react/browser';\nderive(fn())`;
-    const ast = parser.parse(code, {
-      sourceType: 'module',
-      plugins: ['typescript'],
-    });
-    traverse(ast, {
-      ExpressionStatement(path) {
-        const expr = path.get('expression') as NodePath<t.Expression>;
-        const parts = flattenExpressionToParts(expr.node, expr);
-        expect(parts).toHaveLength(1);
-        expect(parts[0].type).toBe('derive');
-        path.stop();
-      },
-    });
-  });
-});
-
-describe('mergeAdjacentStaticParts', () => {
-  it('merges consecutive static parts', () => {
-    const parts = [
-      { type: 'static' as const, value: 'A' },
-      { type: 'static' as const, value: 'B' },
-      { type: 'static' as const, value: 'C' },
-    ];
-    const merged = mergeAdjacentStaticParts(parts);
-    expect(merged).toEqual([{ type: 'static', value: 'ABC' }]);
-  });
-
-  it('does not merge non-adjacent static parts', () => {
-    const node = t.identifier('x');
-    const parts = [
-      { type: 'static' as const, value: 'A' },
-      { type: 'dynamic' as const, node },
-      { type: 'static' as const, value: 'B' },
-    ];
-    const merged = mergeAdjacentStaticParts(parts);
-    expect(merged).toHaveLength(3);
-    expect((merged[0] as any).value).toBe('A');
-    expect(merged[1].type).toBe('dynamic');
-    expect((merged[2] as any).value).toBe('B');
-  });
-});
-
-describe('buildTransformResult', () => {
-  it('returns StringLiteral for all-static parts', () => {
-    const parts = [{ type: 'static' as const, content: 'hello world' }];
-    const { message, variables } = buildTransformResult(parts);
-    expect(t.isStringLiteral(message)).toBe(true);
-    expect((message as t.StringLiteral).value).toBe('hello world');
-    expect(variables).toBeNull();
-  });
-
-  it('returns StringLiteral with placeholders for dynamic parts (no derive)', () => {
-    const node = t.identifier('name');
-    const parts = [
-      { type: 'static' as const, content: 'Hello, ' },
-      { type: 'dynamic' as const, content: node as t.Expression },
-      { type: 'static' as const, content: '!' },
-    ];
-    const { message, variables } = buildTransformResult(parts);
-    expect(t.isStringLiteral(message)).toBe(true);
-    expect((message as t.StringLiteral).value).toBe('Hello, {0}!');
-    expect(variables).not.toBeNull();
-    expect(variables!.properties).toHaveLength(1);
-  });
-
-  it('preserves derive expressions in TemplateLiteral', () => {
-    const deriveNode = t.callExpression(t.identifier('derive'), [
-      t.callExpression(t.identifier('fn'), []),
-    ]);
-    const parts = [
-      { type: 'static' as const, content: 'Hello ' },
-      { type: 'derive' as const, content: deriveNode as t.Expression },
-    ];
-    const { message, variables } = buildTransformResult(parts);
-    expect(t.isTemplateLiteral(message)).toBe(true);
-    const tl = message as t.TemplateLiteral;
-    expect(tl.expressions).toHaveLength(1);
-    expect(tl.expressions[0]).toBe(deriveNode);
-    expect(tl.quasis).toHaveLength(2);
-    expect(tl.quasis[0].value.raw).toBe('Hello ');
-    expect(tl.quasis[1].value.raw).toBe('');
-    expect(variables).toBeNull();
-  });
-});
